Extract to presenter
