--- app/presenters/cohorts/show_presenter.rb
@@ -1,2 +1,0 @@
-module Cohorts
-end
